backend/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
python-dotenv
pydantic-settings
langchain
langchain-openai
langchain-community
pytest
